requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
importlib-metadata==6.7.0
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
pluggy==1.2.0
py==1.11.0
PySocks==1.7.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-selenium==4.0.1
pytest-variables==3.0.0
requests==2.31.0
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==8.2.3
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
wsproto==1.2.0
zipp==3.15.0
